requirements.txt
discord.py~=2.4.0
SQLAlchemy~=2.0.37
dateparser~=1.2.0